refactor: Move RODict implementation to types.py
- Move RODict and rodict() to dedicated types.py module
- Update __init__.py to import from types module
- Keep tests passing with new module structure
