+++ src/rodict/__init__.py
@@ -1,13 +1,3 @@
-from typing import Any, Mapping, MutableMapping, TypeAlias, TypeVar, cast
-
-K = TypeVar("K")
-V = TypeVar("V")
-
-RODict: TypeAlias = Mapping[K, V] | dict[K, V] | MutableMapping[K, V]
-
-
-def rodict(val: RODict[Any, Any]) -> dict[Any, Any]:
-    return cast(dict[Any, Any], val)
-
+from rodict.types import RODict, rodict
 
 __all__ = ["RODict", "rodict"]
